backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pymysql==1.1.0
influxdb-client==1.40.0
langchain==0.1.9
langchain-core==0.1.27
langchain-openai==0.0.8
langchain-community==0.0.24
requests==2.31.0
aiohttp==3.9.3
httpx==0.27.0